rename: Notice 컴포넌트 modal 폴더로 이동
