+++ src/app/pockets/form/letter/page.tsx
@@ -2,11 +2,11 @@
 
 import React, { useState } from "react";
 import { useRouter } from "next/navigation";
-import Notice from "../../../../components/notice";
-import usePocketStore from "../../../store/usePocket";
-import CardList from "../../../../utils/images/cardList";
-import { cardData } from "../../../../utils/images/cardNames";
-import { submitCustomQuestion } from "../../../../api/api-form";
+import Notice from "@/components/modal/notice";
+import usePocketStore from "@/app/store/usePocket";
+import CardList from "@/utils/images/cardList";
+import { cardData } from "@/utils/images/cardNames";
+import { submitCustomQuestion } from "@/api/api-form";
 import Header from "@/components/header/header";
 import BackButton from "@/components/header/back-button";
 import useModifiedStore from "@/app/store/modifiedStore";
