+++ src/app/pockets/form/page.tsx
@@ -3,14 +3,11 @@
 import React, { useEffect, useRef, useState } from "react";
 import { useRouter, useSearchParams } from "next/navigation";
 import { FaPencilAlt } from "react-icons/fa";
-import usePocketStore from "../../store/usePocket";
-import useModifiedStore from "../../../app/store/modifiedStore";
-import Notice from "../../../components/notice";
-import {
-  fetchRandomQuestion,
-  submitCustomQuestion,
-} from "../../../api/api-form";
-import { postEdit } from "../../../api/api-postEdit";
+import usePocketStore from "@/app/store/usePocket";
+import useModifiedStore from "@/app/store/modifiedStore";
+import Notice from "@/components/modal/notice";
+import { fetchRandomQuestion, submitCustomQuestion } from "@/api/api-form";
+import { postEdit } from "@/api/api-postEdit";
 import Header from "@/components/header/header";
 import BackButton from "@/components/header/back-button";
 import { useInputLimit } from "@/hooks/useInputLimit";
