+++ src/app/pockets/page.tsx
@@ -2,10 +2,10 @@
 
 import React, { useCallback, useEffect, useRef, useState } from "react";
 import { useRouter } from "next/navigation";
-import usePocketStore from "../store/usePocket";
-import { fetchLuckyPouches } from "../../api/api-form";
-import Notice from "../../components/notice";
-import { validatePouches, Pouch } from "../../utils/validation/validationPouch";
+import usePocketStore from "@/app/store/usePocket";
+import { fetchLuckyPouches } from "@/api/api-form";
+import Notice from "@/components/modal/notice";
+import { validatePouches, Pouch } from "@/utils/validation/validationPouch";
 import { getPouch } from "@/utils/images/domain";
 import Header from "@/components/header/header";
 import Logo from "@/components/header/logo";
