+++ src/components/modal/notice.tsx
